Move exceptions from network.netty.exception to network.exceptions
