+++ ethereum/p2p/src/main/java/tech/pegasys/pantheon/ethereum/p2p/network/exceptions/BreachOfProtocolException.java
@@ -10,13 +10,11 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.ethereum.p2p.network.netty.exceptions;
+package tech.pegasys.pantheon.ethereum.p2p.network.exceptions;
 
-import tech.pegasys.pantheon.ethereum.p2p.wire.messages.DisconnectMessage.DisconnectReason;
+public class BreachOfProtocolException extends RuntimeException {
 
-public class PeerDisconnectedException extends RuntimeException {
-
-  public PeerDisconnectedException(final DisconnectReason reason) {
-    super("Peer disconnected for reason: " + reason.toString());
+  public BreachOfProtocolException(final String message) {
+    super(message);
   }
 }
