+++ ethereum/p2p/src/main/java/tech/pegasys/pantheon/ethereum/p2p/network/exceptions/IncompatiblePeerException.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.ethereum.p2p.network.netty.exceptions;
+package tech.pegasys.pantheon.ethereum.p2p.network.exceptions;
 
 public class IncompatiblePeerException extends RuntimeException {
