+++ ethereum/p2p/src/main/java/tech/pegasys/pantheon/ethereum/p2p/network/exceptions/PeerDisconnectedException.java
@@ -10,11 +10,13 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.ethereum.p2p.network.netty.exceptions;
+package tech.pegasys.pantheon.ethereum.p2p.network.exceptions;
 
-public class BreachOfProtocolException extends RuntimeException {
+import tech.pegasys.pantheon.ethereum.p2p.wire.messages.DisconnectMessage.DisconnectReason;
 
-  public BreachOfProtocolException(final String message) {
-    super(message);
+public class PeerDisconnectedException extends RuntimeException {
+
+  public PeerDisconnectedException(final DisconnectReason reason) {
+    super("Peer disconnected for reason: " + reason.toString());
   }
 }
