+++ ethereum/p2p/src/test/java/tech/pegasys/pantheon/ethereum/p2p/network/netty/DeFramerTest.java
@@ -24,9 +24,9 @@ import static org.mockito.Mockito.when;
 
 import tech.pegasys.pantheon.ethereum.p2p.api.MessageData;
 import tech.pegasys.pantheon.ethereum.p2p.api.PeerConnection;
-import tech.pegasys.pantheon.ethereum.p2p.network.netty.exceptions.BreachOfProtocolException;
-import tech.pegasys.pantheon.ethereum.p2p.network.netty.exceptions.IncompatiblePeerException;
-import tech.pegasys.pantheon.ethereum.p2p.network.netty.exceptions.PeerDisconnectedException;
+import tech.pegasys.pantheon.ethereum.p2p.network.exceptions.BreachOfProtocolException;
+import tech.pegasys.pantheon.ethereum.p2p.network.exceptions.IncompatiblePeerException;
+import tech.pegasys.pantheon.ethereum.p2p.network.exceptions.PeerDisconnectedException;
 import tech.pegasys.pantheon.ethereum.p2p.network.netty.testhelpers.NettyMocks;
 import tech.pegasys.pantheon.ethereum.p2p.network.netty.testhelpers.SubProtocolMock;
 import tech.pegasys.pantheon.ethereum.p2p.peers.Peer;
